fix: OTP registration crash and unverified-email bug
Two independent bugs, both surfaced by writing the first real test
coverage for the OTP-gated registration flow:

- config/otp.php: env('OTP_EXPIRES_MINUTES', 5) returns a string once the
  var is actually set in .env (the int default only applies when it's
  fully absent). Carbon::addMinutes() strictly type-checks its argument,
  so every registration attempt crashed with a 500. Cast all three otp.*
  env values to int at the config layer.

- OtpController::verify(): email_verified_at was passed into
  User::create([...]), but that column is in User::$guarded (correctly,
  to block mass-assignment from forms) — so it was silently dropped.
  Every OTP-verified user ended up with email_verified_at still null,
  meaning Laravel's `verified` middleware would immediately bounce them
  back to the email-verification prompt right after they'd already
  proven their identity via OTP. Now set explicitly post-create.

Also replaces tests/Feature/Auth/RegistrationTest.php's stale assertions
(it tested the original single-step Breeze flow — 200 render, immediate
auth — which the app deliberately moved away from) with assertions
matching the real redirect-to-OTP behavior, and adds
tests/Feature/Auth/OtpVerificationTest.php covering the OTP-verification
step itself, which previously had zero test coverage.

+++ config/otp.php
@@ -13,17 +13,17 @@ return [
     /**
      * OTP code length (number of characters).
      */
-    'length' => env('OTP_LENGTH', 8),
+    'length' => (int) env('OTP_LENGTH', 8),
 
     /**
      * OTP expiration time in minutes.
      */
-    'expires_minutes' => env('OTP_EXPIRES_MINUTES', 5),
+    'expires_minutes' => (int) env('OTP_EXPIRES_MINUTES', 5),
 
     /**
      * Maximum OTP verification attempts before lockout.
      */
-    'max_attempts' => env('OTP_MAX_ATTEMPTS', 3),
+    'max_attempts' => (int) env('OTP_MAX_ATTEMPTS', 3),
 
     /**
      * Allowed characters for OTP generation.
